Split remaining code quality hotspots
