+++ shared/desktop-clipboard-contracts.ts
@@ -1,0 +1,9 @@
+export type DesktopClipboardSnapshot = {
+  formats: string[];
+  valuesByFormat: Record<string, string>;
+};
+
+export type DesktopClipboardFilePaths = {
+  filePaths: string[];
+  text: string | null;
+};
