+++ shared/desktop-data-contracts.ts
@@ -1,453 +1,8 @@
-export type Thread = {
-  id: string;
-  title: string;
-  age: string;
-  lastModifiedMs?: number;
-  sessionPath?: string;
-  summary?: string;
-  running?: boolean;
-  unread?: boolean;
-  pinned?: boolean;
-};
-
-export type InboxThread = {
-  threadId: string;
-  title: string;
-  projectId: string;
-  projectName: string;
-  sessionPath: string;
-  age: string;
-  lastActivityMs?: number;
-  prompt: string | null;
-  content: string[];
-  preview: string | null;
-  running: boolean;
-  unread: boolean;
-};
-
-export type Project = {
-  id: string;
-  resolvedId?: string;
-  name: string;
-  threads: Thread[];
-  latestModifiedMs?: number;
-  pinned?: boolean;
-  collapsed?: boolean;
-  threadsLoaded?: boolean;
-  threadCount?: number;
-  repoOriginUrl?: string | null;
-  repoOriginChecked?: boolean;
-};
-
-export type ProjectImportCandidate = {
-  projectId: string;
-  name: string;
-  isGitRepo: boolean;
-  hasOrigin: boolean;
-  originUrl: string | null;
-  alreadyImported: boolean;
-};
-
-export type ArchivedThread = {
-  id: string;
-  title: string;
-  age: string;
-  projectId: string;
-  projectName: string;
-  sessionPath: string;
-};
-
-export type ProseMessage = {
-  id: string;
-  role: "assistant" | "user";
-  format?: "prose" | "list";
-  content: string[];
-  thinkingContent?: string[];
-  thinkingHeaders?: string[];
-  thinkingRedacted?: boolean;
-};
-
-export type ToolResultMessage = {
-  id: string;
-  role: "toolResult";
-  toolName: string;
-  content: string[];
-  isError: boolean;
-};
-
-export type BashExecutionMessage = {
-  id: string;
-  role: "bashExecution";
-  command: string;
-  output: string[];
-  exitCode: number | null;
-  cancelled: boolean;
-  truncated: boolean;
-};
-
-export type CustomThreadMessage = {
-  id: string;
-  role: "custom";
-  customType: string;
-  content: string[];
-};
-
-export type SummaryThreadMessage = {
-  id: string;
-  role: "branchSummary" | "compactionSummary";
-  content: string[];
-};
-
-export type Message =
-  | ProseMessage
-  | ToolResultMessage
-  | BashExecutionMessage
-  | CustomThreadMessage
-  | SummaryThreadMessage;
-
-export type ComposerThinkingLevel = "off" | "minimal" | "low" | "medium" | "high" | "xhigh";
-
-export type ComposerStreamingBehavior = "steer" | "followUp" | "stop";
-
-export type ComposerQueuedPrompt = {
-  id: string;
-  mode: Exclude<ComposerStreamingBehavior, "stop">;
-  queueIndex: number;
-  queueSnapshotKey: string;
-  text: string;
-};
-
-export type ComposerModel = {
-  provider: string;
-  id: string;
-  name: string;
-  reasoning: boolean;
-  input: Array<"text" | "image">;
-};
-
-export type ComposerState = {
-  currentModel: ComposerModel | null;
-  availableModels: ComposerModel[];
-  currentThinkingLevel: ComposerThinkingLevel;
-  availableThinkingLevels: ComposerThinkingLevel[];
-  queuedPrompts: ComposerQueuedPrompt[];
-};
-
-export type ComposerAttachment = {
-  path: string;
-  name: string;
-  kind: "directory" | "text" | "image";
-};
-
-export type DesktopClipboardSnapshot = {
-  formats: string[];
-  valuesByFormat: Record<string, string>;
-};
-
-export type DesktopClipboardFilePaths = {
-  filePaths: string[];
-  text: string | null;
-};
-
-export type DictationState = {
-  available: boolean;
-  reason: "missing-model" | "runtime-error" | "unsupported-platform" | null;
-  runtime: "sherpa-onnx-node" | null;
-  modelDirectory: string | null;
-  modelId: string | null;
-  language: string | null;
-  error: string | null;
-};
-
-export type DictationModelId = "tiny.en" | "base.en" | "small.en";
-
-export type DictationModelSummary = {
-  id: DictationModelId;
-  name: string;
-  description: string;
-  downloadSizeBytes: number;
-  downloadSizeLabel: string;
-  installed: boolean;
-  managed: boolean;
-  selected: boolean;
-};
-
-export type DictationModelInstallResult = {
-  ok: boolean;
-  modelId: DictationModelId;
-  error: string | null;
-};
-
-export type DictationModelRemoveResult = {
-  ok: boolean;
-  modelId: DictationModelId;
-  error: string | null;
-};
-
-export type DictationTranscriptionRequest = {
-  audioBase64: string;
-  sampleRate: number;
-  language?: string | null;
-};
-
-export type DictationTranscriptionResult = {
-  ok: boolean;
-  text: string;
-  runtime: "sherpa-onnx-node" | null;
-  modelDirectory: string | null;
-  modelId: string | null;
-  language: string | null;
-  error: string | null;
-};
-
-export type PiPackageCatalogItem = {
-  name: string;
-  version: string;
-  description: string | null;
-  keywords: string[];
-  monthlyDownloads: number;
-  weeklyDownloads: number;
-  searchScore: number;
-  publishedAt: string;
-  updatedAt: string;
-  npmUrl: string;
-  homepageUrl: string | null;
-  repositoryUrl: string | null;
-  source: string;
-  identityKey: string;
-};
-
-export type PiPackageCatalogPage = {
-  query: string;
-  sort: "monthlyDownloads-desc";
-  total: number;
-  nextCursor: number | null;
-  items: PiPackageCatalogItem[];
-};
-
-export type PiConfiguredPackageType = "npm" | "git" | "local";
-
-export type PiConfiguredPackage = {
-  resourceKind: "package" | "extension";
-  source: string;
-  identityKey: string;
-  displayName: string;
-  type: PiConfiguredPackageType;
-  scope: "user" | "project";
-  filtered: boolean;
-  installedPath: string | null;
-  settingsPath: string | null;
-};
-
-export type PiPackageMutationResult = {
-  source: string;
-  normalizedSource: string;
-  configuredPackages: PiConfiguredPackage[];
-};
-
-export type PiSkillCatalogItem = {
-  id: string;
-  skillId: string;
-  name: string;
-  source: string;
-  installs: number;
-  description: string | null;
-  url: string;
-  sourceUrl: string;
-  identityKey: string;
-};
-
-export type PiSkillCatalogPage = {
-  query: string;
-  total: number;
-  items: PiSkillCatalogItem[];
-};
-
-export type PiConfiguredSkill = {
-  source: string;
-  identityKey: string;
-  displayName: string;
-  description: string | null;
-  scope: "user" | "project";
-  provenance: "skills.sh" | "local";
-  installedPath: string;
-  skillFilePath: string;
-  sourceRepo: string | null;
-  sourceUrl: string | null;
-};
-
-export type PiSkillMutationResult = {
-  source: string;
-  normalizedSource: string;
-  configuredSkills: PiConfiguredSkill[];
-};
-
-export type SkillCreatorSessionMessage = {
-  id: string;
-  role: "assistant" | "user";
-  content: string;
-};
-
-export type SkillCreatorSessionState = {
-  sessionId: string;
-  messages: SkillCreatorSessionMessage[];
-  latestResponse: string | null;
-  createdSkillPath: string | null;
-};
-
-export type ComposerFilePickerEntry = {
-  path: string;
-  name: string;
-  kind: "directory" | "text" | "image";
-};
-
-export type ComposerFilePickerState = {
-  homePath: string;
-  rootPath: string;
-  currentPath: string;
-  parentPath: string | null;
-  entries: ComposerFilePickerEntry[];
-};
-
-export type ModelSelection = {
-  provider: string;
-  id: string;
-};
-
-export type ProjectDeletionMode = "pi-only" | "full-clean";
-
-export type AppSettings = {
-  gitCommitMessageModel: ModelSelection | null;
-  skillCreatorModel: ModelSelection | null;
-  composerStreamingBehavior: ComposerStreamingBehavior;
-  dictationModelId: DictationModelId | null;
-  dictationMaxDurationSeconds: number;
-  showDictationButton: boolean;
-  favoriteFolders: string[];
-  projectImportState: boolean | null;
-  preferredProjectLocation: string | null;
-  initializeGitOnProjectCreate: boolean;
-  projectDeletionMode: ProjectDeletionMode;
-  useAgentsSkillsPaths: boolean;
-  piTuiTakeover: boolean;
-};
-
-export type ComposerStateRequest = {
-  projectId?: string | null;
-  sessionPath?: string | null;
-};
-
-export type ProjectGitState = {
-  projectId: string;
-  isGitRepo: boolean;
-  branch: string | null;
-  fileCount: number;
-  stagedFileCount: number;
-  unstagedFileCount: number;
-  insertions: number;
-  deletions: number;
-  hasOrigin: boolean;
-  originName: string | null;
-  originUrl: string | null;
-};
-
-export type ProjectDiffBaseline =
-  | { kind: "head" }
-  | { kind: "previous" }
-  | { kind: "last-opened"; rev: string; capturedAt?: string | null }
-  | { kind: "yesterday" }
-  | { kind: "main-branch" }
-  | { kind: "dev-branch" }
-  | { kind: "commit"; sha: string };
-
-export type ProjectDiffResolvedBaseline = {
-  kind: ProjectDiffBaseline["kind"];
-  rev: string;
-  label: string;
-  commitSha: string | null;
-  shortSha: string | null;
-  subject: string | null;
-  committedAt: string | null;
-  capturedAt: string | null;
-};
-
-export type ProjectCommitEntry = {
-  sha: string;
-  shortSha: string;
-  subject: string;
-  authorName: string;
-  authorEmail: string;
-  authoredAt: string;
-  committedAt: string;
-  decorations: string[];
-  isHead: boolean;
-};
-
-export type ProjectDiffResult = {
-  projectId: string;
-  diff: string;
-  fileCount: number;
-  insertions: number;
-  deletions: number;
-  baseline: ProjectDiffBaseline;
-  resolvedBaseline: ProjectDiffResolvedBaseline;
-};
-
-export type ProjectDiffStatsResult = {
-  projectId: string;
-  fileCount: number;
-  insertions: number;
-  deletions: number;
-  baseline: ProjectDiffBaseline;
-  resolvedBaseline: ProjectDiffResolvedBaseline;
-};
-
-export type ShellState = {
-  platform: string;
-  mockMode: boolean;
-  productName: string;
-  cwd: string;
-  resolvedCwd?: string;
-  agentDir: string;
-  sessionDir: string;
-  projects: Project[];
-  appSettings: AppSettings;
-  availableHosts: string[];
-  composer: ComposerState;
-};
-
-export type ThreadData = {
-  sessionPath: string;
-  title: string;
-  messages: Message[];
-  previousMessageCount: number;
-  isStreaming: boolean;
-};
-
-export type DesktopEvent =
-  | {
-      type: "shell-state-refresh";
-    }
-  | {
-      type: "dictation-download-log";
-      modelId: DictationModelId;
-      message: string;
-      at: string;
-      done: boolean;
-      isError: boolean;
-    }
-  | {
-      type: "thread-update";
-      reason: "start" | "update" | "end" | "external";
-      projectId: string;
-      threadId: string;
-      sessionPath: string;
-      thread: ThreadData;
-      composer: ComposerState | null;
-    }
-  | {
-      type: "composer-update";
-      projectId: string | null;
-      sessionPath: string | null;
-      composer: ComposerState;
-    };
+export * from "./desktop-clipboard-contracts";
+export * from "./desktop-composer-contracts";
+export * from "./desktop-dictation-contracts";
+export * from "./desktop-event-contracts";
+export * from "./desktop-package-contracts";
+export * from "./desktop-project-git-contracts";
+export * from "./desktop-settings-contracts";
+export * from "./desktop-thread-contracts";
